pyproject.toml
[tool.poetry]
name = "rest-api-project"
version = "0.1.0"
description = "REST API solution for Adastra"
authors = ["nipunaupeksha <nipuna.upeksha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.17"
bcrypt = "3.2.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-cov = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["./rest_api_project"]
